ci: fix build on Windows (reference to value converters)
