+++ src/SamplesApp/UITests.Shared/Windows_Devices/PedometerTests.xaml
@@ -1,14 +1,13 @@
 ﻿<UserControl
-    x:Class="UITests.Shared.Windows_Devices.PedometerTests"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:UITests.Shared.Windows_Devices"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	xmlns:valueconverters="using:UITests.Shared.ValueConverters"
+	x:Class="UITests.Shared.Windows_Devices.PedometerTests"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	xmlns:valueconverters="using:UITests.ValueConverters"
 	mc:Ignorable="d"
-    d:DesignHeight="300"
-    d:DesignWidth="400">
+	d:DesignHeight="300"
+	d:DesignWidth="400">
 
 	<UserControl.Resources>
 		<valueconverters:BoolNegationValueConverter x:Key="BoolNegation" />
