+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Button/RadioButton_Pressed.xaml
@@ -1,14 +1,16 @@
-﻿<UserControl x:Class="UITests.Shared.Windows_UI_Xaml_Controls.Button.RadioButton_Pressed"
-			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:local="using:UITests.Shared.Windows_UI_Xaml_Controls.Button"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 mc:Ignorable="d"
-			 d:DesignHeight="300"
-			 d:DesignWidth="400">
+﻿<UserControl
+	x:Class="UITests.Shared.Windows_UI_Xaml_Controls.Button.RadioButton_Pressed"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:valueconverters="using:UITests.ValueConverters"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	mc:Ignorable="d"
+	d:DesignHeight="300"
+	d:DesignWidth="400">
+
 	<UserControl.Resources>
-		<local:FromNullableBoolToCustomValueConverter x:Key="BoolToIconStyle"
+		<valueconverters:FromNullableBoolToCustomValueConverter x:Key="BoolToIconStyle"
 													  NullOrFalseValue="{StaticResource GridUnselectedStyle}"
 													  TrueValue="{StaticResource GridSelectedStyle}" />
 		<Style x:Key="RadioButtonStyle"
