+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Controls/Button/RadioButton_Pressed.xaml.cs
@@ -27,28 +27,4 @@ namespace UITests.Shared.Windows_UI_Xaml_Controls.Button
             this.InitializeComponent();
         }
 	}
-
-
-	public class FromNullableBoolToCustomValueConverter : IValueConverter
-	{
-		public object NullOrFalseValue { get; set; }
-		public object TrueValue { get; set; }
-
-		public object Convert(object value, Type targetType, object parameter, string language)
-		{
-			if (value == null || !System.Convert.ToBoolean(value, CultureInfo.InvariantCulture))
-			{
-				return NullOrFalseValue;
-			}
-			else
-			{
-				return TrueValue;
-			}
-		}
-
-		public object ConvertBack(object value, Type targetType, object parameter, string language)
-		{
-			throw new NotImplementedException();
-		}
-	}
 }
